backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
bcrypt==4.1.3
billiard==4.2.4
black==24.10.0
boto3==1.35.33
botocore==1.35.99
celery==5.4.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.12.0
cryptography==43.0.3
dj-database-url==3.0.1
Django==5.2.8
django-auditlog==3.0.0
django-cors-headers==4.4.0
django-filter==23.5
django-ratelimit==4.0.0
django-redis==5.4.0
django-sendgrid-v5==0.8.1
django-storages==1.14.5
djangorestframework==3.16.1
djangorestframework-api-key==2.1.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.8
ecdsa==0.19.1
et_xmlfile==2.0.0
flake8==7.1.1
future==1.0.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.5
openpyxl==3.1.4
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycodestyle==2.12.1
pycparser==2.23
pyflakes==3.2.0
PyJWT==2.10.1
pypng==0.20220715.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-http-client==3.3.7
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.3
qrcode==7.4.2
redis==5.0.7
referencing==0.37.0
reportlab==4.2.2
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.10.4
sendgrid==6.12.4
sentry-sdk==2.15.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.4
